/api/chat: Implement basic timing data

+++ server/ollama-types/response-types.go
@@ -9,10 +9,10 @@ type ChatResponse struct {
 
 type ChatFinalResponse struct {
 	ChatResponse
-	TotalDuration      int `json:"total_duration"`
-	LoadDuration       int `json:"load_duration"`
-	PromptEvalCount    int `json:"prompt_eval_count"`
-	PromptEvalDuration int `json:"prompt_eval_duration"`
-	EvalCount          int `json:"eval_count"`
-	EvalDuration       int `json:"eval_duration"`
+	TotalDuration      int64 `json:"total_duration"`
+	LoadDuration       int64 `json:"load_duration"`
+	PromptEvalCount    int64 `json:"prompt_eval_count"`
+	PromptEvalDuration int64 `json:"prompt_eval_duration"`
+	EvalCount          int64 `json:"eval_count"`
+	EvalDuration       int64 `json:"eval_duration"`
 }
